modules/persons_ws/requirements-runtime.txt
grpcio==1.54.0
grpcio-tools==1.54.0
SQLAlchemy==2.0.10
psycopg2==2.9.6